fix code

+++ backend/src/main/java/com/trello/service/impl/WorkspaceServiceImpl.java
@@ -1,6 +1,8 @@
 package com.trello.service.impl;
 
+import com.trello.domain.Task;
 import com.trello.domain.Workspace;
+import com.trello.repository.TaskRepository;
 import com.trello.repository.WorkspaceRepository;
 import com.trello.service.WorkspaceService;
 import com.trello.service.dto.WorkspaceDTO;
@@ -10,6 +12,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -20,10 +23,11 @@ public class WorkspaceServiceImpl implements WorkspaceService {
 
     private final WorkspaceRepository workspaceRepository;
     private final WorkspaceMapper workspaceMapper;
-
-    public WorkspaceServiceImpl(WorkspaceRepository workspaceRepository, WorkspaceMapper workspaceMapper) {
+    private final TaskRepository taskRepository;
+    public WorkspaceServiceImpl(WorkspaceRepository workspaceRepository, WorkspaceMapper workspaceMapper, TaskRepository taskRepository) {
         this.workspaceRepository = workspaceRepository;
         this.workspaceMapper = workspaceMapper;
+        this.taskRepository = taskRepository;
     }
 
     @Override
@@ -34,11 +38,14 @@ public class WorkspaceServiceImpl implements WorkspaceService {
 
     @Override
     public WorkspaceDTO createWorkspace(WorkspaceDTO workspaceDTO) {
+        if(workspaceDTO.getId() != null) {
+            throw new IllegalArgumentException("workspace id must be null");
+        }
         if(workspaceDTO.getName().isBlank())
             throw new IllegalArgumentException("Workspace name cannot be blank");
         if(workspaceDTO.getName().length() > 255)
             throw new IllegalArgumentException("Workspace name cannot be longer than 255 characters");
-
+        workspaceDTO.setCreatedAt(LocalDateTime.now());
         Workspace savedWorkspace = workspaceRepository.save(
                 workspaceMapper.toEntity(workspaceDTO));
         return workspaceMapper.toDTO(savedWorkspace);
@@ -68,6 +75,8 @@ public class WorkspaceServiceImpl implements WorkspaceService {
             throw new IllegalArgumentException("Workspace is not exist");
 
         workspaceDTO.setId(workspaceId);
+        workspaceDTO.setUpdatedAt(LocalDateTime.now());
+        workspaceDTO.setCreatedAt(workspace.get().getCreatedAt());
         workspaceRepository.save(workspaceMapper.toEntity(workspaceDTO));
     }
 
@@ -75,6 +84,10 @@ public class WorkspaceServiceImpl implements WorkspaceService {
     public void deleteById(Long workspaceId) {
         if(workspaceId == null || !workspaceRepository.existsById(workspaceId))
             throw new IllegalArgumentException("Workspace is not exist");
+        Optional<Task> task = taskRepository.findByWorkspaceId(workspaceId);
+        if(task.isPresent()){
+            throw new IllegalArgumentException("task found with id " + workspaceId);
+        }
         workspaceRepository.deleteById(workspaceId);
     }
 }
